% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fchl_correctoR.R
\name{correct_fchl}
\alias{correct_fchl}
\title{Temperature correction for (some) chlorophyll fluorometers}
\usage{
correct_fchl(
  raw_fchl = NULL,
  temp_c = NULL,
  instr = "EXO2",
  bias_corr = NULL,
  na_rm = TRUE
)
}
\arguments{
\item{raw_fchl}{Vector of numbers}

\item{temp_c}{Vector of numbers}

\item{instr}{One of <X, Y, Z>}

\item{na_rm}{Whether \code{NA}s should be removed. Defaults to \code{TRUE}}
}
\value{
Corrected chlorophyll concentration data using 'temp_c' for the numbers provided to 'raw_fchl'
}
\description{
Requires temperature (celsius) and raw chlorophyll fluorescence data (ug/L). Applies Watras (2017) correction using 25 degrees celcius as the reference temperature and Richardson (2025) bias corrections. This should be applied to YSI 6600, YSI EXO, and Seabird WETStar data.
}
\examples{
# Temperature correction (for 6600, EXO2, or WS)

}
